Rename `meta::stack` to `meta::type_sequence`.
Originally, the class was designed to serve as a symbol stack for the LL(1) parsing algorithm. Over the course of developing this project, it has been used in places which do not require stack-like semantics, like AST inversion or capture name specification. This new name better reflects the design goal of the data structure.